Alert user if more specific audit_*() summaries are available

(Note: Ignore this function if your audit() method calls audit_cols_minimal().)

Call check_audit_special() within an audit() method for a consistency test mapper function, such as audit.scr_grim_map(). It checks if the input data frame was the product of a function produced by function_map_seq() or function_map_total_n().

If so, the function issues a gentle alert to the user that points to audit_seq() or audit_total_n(), respectively.

Usage

check_audit_special(data, name_test)

Arguments

data

The audit() method's input data frame.

name_test

String (length 1). Short, plain-text name of the consistency test, such as "GRIM".

Value

No return value. Might print an alert.

See also