Count duplicate values

duplicate_count() returns a frequency table. When searching a data frame, it includes values from all columns for each frequency count.

This function is a blunt tool designed for initial data checking. Don't put too much weight on its results.

For summary statistics, call audit() on the results.

duplicate_count(x, numeric_only = TRUE)

Arguments

x
Vector or data frame.

numeric_only
Boolean. If TRUE (the default), and if x is a data frame, the function includes only numeric columns and string columns coercible to numeric. Note: Be careful when setting it to FALSE. This can lead to all kinds of coercion issues.

Value

A tibble with two columns —

value includes all the values from x.

count is the frequency of each value in x, in descending order. The tibble has the scr_dup_count class, which is recognized by the audit() generic.

Details

duplicate_count() is a thin wrapper around janitor::get_dupes(). Use get_dupes() to search for duplicate rows.

The function is not too informative if the values have few characters.

Summaries with audit()

There is an S3 method for the audit() generic, so you can call audit() following duplicate_count() to get summary statistics. These are mostly self-explaining, but count_max and count_min only directly apply to count and display their respective value numbers, not the minimal and maximal value numbers.

See also

Examples

# Count duplicate values...
BJsales %>%
  duplicate_count()
#> # A tibble: 22 × 2
#>    value count
#>    <dbl> <int>
#>  1  209.     4
#>  2  219.     3
#>  3  257.     3
#>  4  258.     3
#>  5  202.     2
#>  6  210.     2
#>  7  211.     2
#>  8  212.     2
#>  9  215.     2
#> 10  216.     2
#> # … with 12 more rows

# ...and compute summaries:
BJsales %>%
  duplicate_count() %>%
  audit()
#> # A tibble: 2 × 7
#>   term  count_max count_min     n   mean     sd median
#>   <chr>     <dbl>     <dbl> <int>  <dbl>  <dbl>  <dbl>
#> 1 value      209.      260.    22 232.   21.3     221.
#> 2 count        4         2     22   2.23  0.528     2