The GRIM test (granularity-related inconsistency of means)

grim() checks if a reported mean value of integer data is mathematically consistent with the reported sample size and the number of items that compose the mean value.

Set percent to TRUE if x is a percentage. This will convert x to a decimal number and adjust the decimal count accordingly.

The function is vectorized, but it is recommended to use grim_map() for testing multiple cases.

grim(
  x,
  n,
  items = 1,
  percent = FALSE,
  show_rec = FALSE,
  rounding = "up_or_down",
  threshold = NULL,
  symmetric = FALSE,
  tolerance = .Machine$double.eps^0.5
)

Arguments

x
String. The reported mean or percentage value.

n
Integer. The reported sample size.

items
Numeric. The number of items composing x. Default is 1, the most common case.

percent
Boolean. Set percent to TRUE if x is a percentage. This will convert it to a decimal number and adjust the decimal count (i.e., increase it by 2). Default is FALSE.

show_rec
Boolean. If set to TRUE, the output is a matrix that also contains intermediary values from GRIM-testing. This is not recommended; instead, use show_rec in grim_map(). Default is FALSE.

rounding
String. Rounding method or methods to be used for reconstructing the values to which x will be compared. Default is "up_or_down" (from 5). For more options, see Details.

threshold
Numeric. If rounding is set to "up_from", "down_from", or "up_from_or_down_from", set threshold to the number from which the reconstructed values should then be rounded up or down. Otherwise, this argument plays no role. Default is NULL.

symmetric
Boolean. Set symmetric to TRUE if the rounding of negative numbers with "up", "down", "up_from", or "down_from" should mirror that of positive numbers so that their absolute values are always equal. Default is FALSE.

tolerance
Numeric. Tolerance of comparison between x and the possible mean or percentage values. Default is circa 0.000000015 (1.490116e-08), as in dplyr::near().

Value

Boolean. TRUE if x, n, and items are mutually consistent, FALSE if not.

Details

The x values need to be strings because only strings retain trailing zeros, which are as important for the GRIM test as any other decimal digits.

Use restore_zeros() on numeric values (or values that were numeric values at some point) to easily supply the trailing zeros they might once have had. See documentation there.

Browse the source code in the grim.R file. grim() is a vectorized version of the internal grim_scalar() function found there.

The GRIM test (granularity-related inconsistency of means) was devised by ---REFERENCE TO BROWN AND HEATHERS 2017---.

Rounding

Here are the options for the rounding argument. Reconstructed mean or percentage values can be rounded in either of these ways:

Rounded to "even" using base R's own round().

Rounded "up" or "down" from 5. (Note that SAS, SPSS, Stata, Matlab, and Excel round "up" from 5, whereas Python rounds "down" from 5.)

Rounded "up_from" or "down_from" some number, which then needs to be specified via the threshold argument.

Given a "ceiling" or "floor" at the respective decimal place.

Rounded towards zero with "trunc" or away from zero with "anti_trunc". The default, "up_or_down", allows for numbers rounded either "up" or "down" from 5 when GRIM-testing; and likewise for "up_from_or_down_from" and "ceiling_or_floor".

With rounding = "up_or_down", if n is 40 or 80 and x has two decimal places, very few values will test as inconsistent; but note that many will be with either of rounding = "up" and rounding = "down", or indeed with any other rounding method. This is part of a more general pattern: n is 400 or 800 and x has three decimal places, etc.

For more information about these methods, see documentation for round(), round_up(), and round_ceiling(). These include all of the above ways of rounding. How the reconstructed values are rounded can also be calibrated by the threshold and symmetric arguments.

See also

Examples

# A mean of 5.19 is not consistent with a sample size of 28:
grim(x = "5.19", n = 28)    # `x` in quotes!
#>  5.19 
#> FALSE 

# However, it is consistent with a sample size of 32:
grim(x = "5.19", n = 32)
#> 5.19 
#> TRUE 

# For a scale composed of two items:
grim(x = "2.84", n = 16, items = 2)
#> 2.84 
#> TRUE 

# With percentages instead of means -- here, 71%:
grim(x = "71", n = 43, percent = TRUE)
#>    71 
#> FALSE