Visualize GRIM test results

grim_plot() visualizes summary data and their mutual GRIM consistency. Call this function only on a data frame that resulted from a call to grim_map() in which adjust_x was set to 1.

Consistent and inconsistent value pairs from the input data frame are shown in distinctive colors. By default, consistent value pairs are blue and inconsistent ones are red. These and other parameters of the underlying geoms can be controlled via arguments.

The background raster flags every inconsistent value pair in the space defined by the decimals argument. That is only informative for value pairs where the mean/percentage has exactly that many decimal places. The raster is independent of the data, but it follows the rounding specification in the grim_map() call.

grim_plot(
  data = NULL,
  show_raster = TRUE,
  show_zoom = TRUE,
  show_full_range = TRUE,
  n = NULL,
  decimals = 2,
  color_cons = "royalblue1",
  color_incons = "red",
  tile_alpha = 1,
  tile_size = 1.5,
  raster_alpha = 0.5,
  raster_color = "gray50"
)

Arguments

data
Data frame. Result of a call to grim_map().

show_raster
Boolean. If TRUE (the default), the plot has a background raster.

show_zoom
Boolean. Should zoom be added if the number of decimal places is 3 or greater? Default is TRUE.

show_full_range
Boolean. Should the full range of the plot always be shown, regardless of the empirical values? Default is TRUE.

n
Integer. Maximal value on the x-axis. Default is NULL, in which case n becomes 10 ^ decimals (e.g., 100 if decimals is 2).

decimals
Integer. Number of decimal places for which the background raster will be generated. Only 1 and 2 are currently supported. Default is 2.

color_cons, color_incons
Strings. Fill colors of the consistent and inconsistent scatter points. Defaults are "royalblue1" (consistent) and "red" (inconsistent).

tile_alpha, tile_size
Numeric. Further parameters of the scatter points: opacity and, indirectly, size. Defaults are 1 and 1.5.

raster_alpha, raster_color
Numeric and string, respectively. Parameters of the background raster: opacity and fill color. Defaults are 0.5 and "gray50".

Background raster

The background raster flags inconsistent value sets with dark boxes. All other places in the plot denote consistent value sets.

As the raster varies by rounding procedure, it will automatically correspond to the rounding argument specified in the preceding grim_map() call. This works fast because the raster is based on data saved in the package itself, so these data don't need to be generated anew every time the function is called.

With the default, "up_or_down", strikingly few values are flagged as inconsistent for sample sizes 40 and 80 (or 400 and 800, etc.). This effect disappears if rounding is set to any other value. For a list of values that rounding can take, see documentation for grim(), section Rounding.

The 4/8 leniency effect arises because accepting values rounded either up or down is more careful and conservative than any other rounding procedure. In any case, grim_plot() doesn't cause this effect --- it only reveals it.

The number of tiles in the plot, including both the raster and the points spared by it, is \((10 ^ {decimals}) ^ 2\). So it's 10.000 with the default decimals = 2, and one million with decimals = 3. For each greater decimals value (with decimals >= 2), the number of tiles is 100 times the former value!