Percentages and sample sizes for GRIM-testing

A fictional dataset with percentages and sample sizes of flying pigs. It can be used to demonstrate the functionality of grim_map(), particularly its percent argument, and functions building up on it.

Usage

pigs2

Format

A tibble (data frame) with 6 rows and 2 columns. The columns are:

x

String. Percentages.

n

Numeric. Sample sizes.

Value

A tibble (data frame).

See also