Data with duplications

A fictional dataset with observations of flying pigs. It contains multiple duplicates. The dataset can be used to demonstrate the functionality of duplicate_*() functions such as duplicate_count().

Usage

pigs4

Format

A tibble (data frame) with 5 rows and 3 columns, describing various body measures of the fictional pigs. The columns are:

snout

String. Snout width.

tail

String. Tail length.

wings

String. Wingspan.

Value

A tibble (data frame).

See also