Restore trailing zeros

restore_zeros() takes a vector with values that might have lost trailing zeros, most likely from being registered as numeric. It turns each value into a string and adds trailing zeros until the mantissa hits some limit.

The default for that limit is the number of digits in the longest mantissa of the vector's values. The length of the integer part plays no role.

Don't rely on the default limit without checking: The original width could have been larger because the longest extant mantissa might itself have lost trailing zeros.

restore_zeros(x, width = NULL, sep = "\\.")

Arguments

x
Numeric (or string coercible to numeric). Vector of numbers that might have lost trailing zeros.

width
Integer. Number of decimal places the mantissas should have, including the restored zeros. Default is NULL, in which case the number of characters in the longest mantissa will be used instead.

sep
Substring that separates the mantissa from the integer part. Default is "\\.", which renders a decimal point.

Value

A string vector. At least some of the strings will have newly restored zeros, unless all input values had the same number of decimal places.

Details

This function exploits the fact that groups of summary values such as means or percentages are often reported to the same number of decimal places. If such a number is known but values were not entered as strings, trailing zeros will be lost. In this case, restore_zeros() will be helpful to prepare data for consistency testing functions such as grim_map() or debit_map(). Otherwise, it should probably not be used.

Displaying decimal places

You might not see all decimal places of numeric values in a vector, and consequently wonder if restore_zeros(), when applied to the vector, adds too many zeros. That is because displayed numbers, unlike stored numbers, are often rounded.

For a vector x, you can count the characters of the longest mantissa from among its values like this:

x %>% decimal_places() %>% max()

See also

Examples

# By default, the target width is that of
# the longest mantissa:
vec <- c(212, 75.38, 4.9625)
vec %>% restore_zeros()
#> [1] "212.0000" "75.3800"  "4.9625"  

# Alternatively, supply a number via `width`:
vec %>% restore_zeros(width = 6)
#> [1] "212.000000" "75.380000"  "4.962500"