Rank sequence test results

Run this function after generating a sequence with seq_endpoint_df() or seq_distance_df() and testing it with one of scrutiny's mapping functions, such as grim_map(). It will rank the test's consistent and inconsistent results by their positions in the sequence.

seq_test_ranking(x, explain = TRUE)

Arguments

x
Data frame.

explain
If TRUE (the default), results come with an explanation.

Details

The function checks the provenance of the test results and throws a warning if it's not correct.