Split columns by parentheses

Summary statistics are often presented like "2.65 (0.27)". When working with tables from PDF, it can be tedious to separate values before and inside parentheses. split_by_parens() does that automatically. By default, it operates on all columns. Output can optionally be transformed into a longer format by setting .transform to TRUE.

split_by_parens(
  .data,
  ...,
  .keep = FALSE,
  .transform = FALSE,
  .sep = "parens",
  .col1 = "x",
  .col2 = "sd"
)

Arguments

.data

Data frame. Only in split_by_parens().

...

Optionally, name columns from .data. Splitting will then be restricted to these columns. This is useful if not all values contain parentheses.

.keep

Boolean. If set to TRUE, the original columns from .data also appear in the output. Default is FALSE.

.transform

Boolean. If set to TRUE, the output will be transformed to be better accessible for typical follow-up tasks. Default is FALSE.

.sep

String. What to split by. Either "parens", "brackets", or "braces". Default is "parens".

.col1, .col2

Strings. Endings of the two column names that result from splitting a column. Default is "x" for .col1 and "sd" for .col2.

Value

A tibble with string columns.

See also

Examples

# Call `split_by_parens()` on data like these:
pigs <- tibble::tribble(
  ~drone,           ~selfpilot,
  "0.09 (0.21)",    "0.19 (0.13)",
  "0.19 (0.28)",    "0.53 (0.10)",
  "0.62 (0.16)",    "0.50 (0.11)",
  "0.15 (0.35)",    "0.57 (0.16)",
)

pigs
#> # A tibble: 4 × 2
#>   drone       selfpilot  
#>   <chr>       <chr>      
#> 1 0.09 (0.21) 0.19 (0.13)
#> 2 0.19 (0.28) 0.53 (0.10)
#> 3 0.62 (0.16) 0.50 (0.11)
#> 4 0.15 (0.35) 0.57 (0.16)

# Basic usage:
pigs %>%
  split_by_parens()
#> # A tibble: 4 × 4
#>   drone_x drone_sd selfpilot_x selfpilot_sd
#>   <chr>   <chr>    <chr>       <chr>       
#> 1 0.09    0.21     0.19        0.13        
#> 2 0.19    0.28     0.53        0.10        
#> 3 0.62    0.16     0.50        0.11        
#> 4 0.15    0.35     0.57        0.16        

# Name specific columns to only return those:
pigs %>%
  split_by_parens(drone)
#> # A tibble: 4 × 2
#>   drone_x drone_sd
#>   <chr>   <chr>   
#> 1 0.09    0.21    
#> 2 0.19    0.28    
#> 3 0.62    0.16    
#> 4 0.15    0.35    

# Pivot the data into a longer format
# by setting `.transform` to `TRUE`:
pigs %>%
  split_by_parens(.transform = TRUE)
#> # A tibble: 8 × 3
#>   .origin   x     sd   
#>   <chr>     <chr> <chr>
#> 1 drone     0.09  0.21 
#> 2 drone     0.19  0.28 
#> 3 drone     0.62  0.16 
#> 4 drone     0.15  0.35 
#> 5 selfpilot 0.19  0.13 
#> 6 selfpilot 0.53  0.10 
#> 7 selfpilot 0.50  0.11 
#> 8 selfpilot 0.57  0.16 

# Choose different column names or
# name suffixes with `.col1` and `.col2`:
pigs %>%
  split_by_parens(.col1 = "beta", .col2 = "se")
#> # A tibble: 4 × 4
#>   drone_beta drone_se selfpilot_beta selfpilot_se
#>   <chr>      <chr>    <chr>          <chr>       
#> 1 0.09       0.21     0.19           0.13        
#> 2 0.19       0.28     0.53           0.10        
#> 3 0.62       0.16     0.50           0.11        
#> 4 0.15       0.35     0.57           0.16        

pigs %>%
  split_by_parens(.transform = TRUE,
  .col1 = "beta", .col2 = "se")
#> # A tibble: 8 × 3
#>   .origin   beta  se   
#>   <chr>     <chr> <chr>
#> 1 drone     0.09  0.21 
#> 2 drone     0.19  0.28 
#> 3 drone     0.62  0.16 
#> 4 drone     0.15  0.35 
#> 5 selfpilot 0.19  0.13 
#> 6 selfpilot 0.53  0.10 
#> 7 selfpilot 0.50  0.11 
#> 8 selfpilot 0.57  0.16