Summaries of scrutiny objects

audit() is an S3 generic to follow up on those scrutiny functions that run tests on a data frame. It summarizes results of those tests and presents the summaries in a tibble.

Below is a list of functions that return objects with classes for which there are audit() methods. This means you can run audit() on the output returned by any of these functions.

Go to the documentation of any particular function named below to learn about its audit() method.

Function	Class

grim_map()	scr_grim_map

debit_map()	scr_debit_map

duplicate_count()	scr_dup_count

duplicate_detect()	scr_dup_detect

audit(data)

Arguments

data
A data frame with one of the classes named above.

Value

A tibble (data frame) with test summary statistics.

Examples

# For GRIM-testing:
pigs1 %>%
  grim_map() %>%
  audit()
#> # A tibble: 1 × 7
#>   incons_cases all_cases incons_rate mean_grim_ratio incons_to_ratio
#>          <int>     <int>       <dbl>           <dbl>           <dbl>
#> 1            8        12       0.667           0.724           0.921
#> # … with 2 more variables: testable_cases <int>, testable_rate <dbl>

# For detecting duplicates:
pigs4 %>%
  duplicate_detect() %>%
  audit()
#> # A tibble: 3 × 4
#>   variable n_duplicated n_total dup_rate
#>   <chr>           <int>   <int>    <dbl>
#> 1 snout               3       5      0.6
#> 2 tail                1       5      0.2
#> 3 .total              4      10      0.4