Using DEBIT with dispersed inputs

debit_map_seq() applies DEBIT with values surrounding the input values. This provides an easy and powerful way to assess whether small errors in computing or reporting may be responsible for DEBIT-inconsistencies in published statistics.

Usage

debit_map_seq(
  data,
  x = NULL,
  sd = NULL,
  n = NULL,
  var = .var,
  dispersion = .dispersion,
  out_min = .out_min,
  out_max = .out_max,
  include_reported = .include_reported,
  include_consistent = .include_consistent,
  ...
)

Arguments

data

A data frame that debit_map() could take.

x, sd, n

Optionally, specify column names in data as these arguments.

var

String. Names of the columns that will be dispersed. Default is c("x", "sd", "n").

dispersion

Numeric. Sequence with steps up and down from the var inputs. It will be adjusted to these values' decimal levels. For example, with a reported 8.34, the step size is 0.01. Default is 1:5, for five steps up and down.

out_min, out_max

If specified, output will be restricted so that it's not below out_min or above out_max. Defaults are "auto" for out_min, i.e., a minimum of one decimal unit above zero; and NULL for out_max, i.e., no maximum.

include_reported

Boolean. Should the reported values themselves be included in the sequences originating from them? Default is FALSE because this might be redundant and bias the results.

include_consistent

Boolean. Should the function also process consistent cases (from among those reported), not just inconsistent ones? Default is FALSE because the focus should be on clarifying inconsistencies.

...

Arguments passed down to debit_map().

Value

A tibble (data frame) with detailed test results.

Summaries with audit_seq()

You can call audit_seq() following debit_map_seq(). It will return a data frame with these columns:

x, sd, and n are the original inputs, tested for consistency here.

hits_total is the total number of DEBIT-consistent value sets found within the specified dispersion range.

hits_x is the number of DEBIT-consistent value sets found by varying x.

Accordingly with sd and hits_sd as well as n and hits_n.

(Note that any consistent reported cases will be counted by the hits_* columns if both include_reported and include_consistent are set to TRUE.)

diff_x reports the absolute difference between x and the next consistent dispersed value (in dispersion steps, not the actual numeric difference). diff_x_up and diff_x_down report the difference to the next higher or lower consistent value, respectively.

diff_sd, diff_sd_up, and diff_sd_down do the same for sd.

Likewise with diff_n, diff_n_up, and diff_n_down.

Call audit() following audit_seq() to summarize results even further.

Examples

# `debit_map_seq()` can take any input
# that `debit_map()` can take:
pigs3
#> # A tibble: 7 × 3
#>   x     sd        n
#>   <chr> <chr> <dbl>
#> 1 0.53  0.50   1683
#> 2 0.44  0.50   1683
#> 3 0.77  0.42   1683
#> 4 0.19  0.35   1683
#> 5 0.34  0.47   1683
#> 6 0.93  0.25   1683
#> 7 0.12  0.33   1683

# Results from testing the first two rows:
out <- pigs3 %>%
  dplyr::slice(3:4) %>%
  debit_map_seq(include_consistent = TRUE)

out
#> # A tibble: 60 × 13
#>    x     sd        n consistency round…¹ sd_lo…² sd_in…³ sd_up…⁴ sd_in…⁵ x_lower
#>    <chr> <chr> <int> <lgl>       <chr>     <dbl> <lgl>     <dbl> <lgl>     <dbl>
#>  1 0.72  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.715
#>  2 0.73  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.725
#>  3 0.74  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.735
#>  4 0.75  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.745
#>  5 0.76  0.42   1683 TRUE        up_or_…   0.415 TRUE      0.425 TRUE      0.755
#>  6 0.78  0.42   1683 TRUE        up_or_…   0.415 TRUE      0.425 TRUE      0.775
#>  7 0.79  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.785
#>  8 0.80  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.795
#>  9 0.81  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.805
#> 10 0.82  0.42   1683 FALSE       up_or_…   0.415 TRUE      0.425 TRUE      0.815
#> # … with 50 more rows, 3 more variables: x_upper <dbl>, case <int>, var <chr>,
#> #   and abbreviated variable names ¹​rounding, ²​sd_lower, ³​sd_incl_lower,
#> #   ⁴​sd_upper, ⁵​sd_incl_upper

# Case-wise summaries with `audit_seq()`
# can be more important than the raw results:
out %>%
  audit_seq()
#> # A tibble: 2 × 17
#>   x     sd        n consi…¹ hits_…² hits_x hits_sd hits_n diff_x diff_…³ diff_…⁴
#>   <chr> <chr> <int> <lgl>     <int>  <int>   <int>  <int>  <dbl>   <dbl>   <dbl>
#> 1 0.77  0.42   1683 TRUE         12      2       0     10      1       1      -1
#> 2 0.19  0.35   1683 FALSE         4      2       2      0      4      NA      -4
#> # … with 6 more variables: diff_sd <dbl>, diff_sd_up <dbl>, diff_sd_down <dbl>,
#> #   diff_n <dbl>, diff_n_up <dbl>, diff_n_down <dbl>, and abbreviated variable
#> #   names ¹​consistency, ²​hits_total, ³​diff_x_up, ⁴​diff_x_down

# Summarize across cases with `audit()`
# following up on `audit_seq()`:
out %>%
  audit_seq() %>%
  audit() %>%
  suppressWarnings()
#> # A tibble: 13 × 7
#>    term          mean    sd median   min   max na_count
#>    <chr>        <dbl> <dbl>  <dbl> <dbl> <dbl>    <dbl>
#>  1 hits_total     8    5.66    8       4    12        0
#>  2 hits_x         2    0       2       2     2        0
#>  3 hits_sd        1    1.41    1       0     2        0
#>  4 hits_n         5    7.07    5       0    10        0
#>  5 diff_x         2.5  2.12    2.5     1     4        0
#>  6 diff_x_up      1   NA       1       1     1        1
#>  7 diff_x_down   -2.5  2.12   -2.5    -4    -1        0
#>  8 diff_sd        4   NA       4       4     4        1
#>  9 diff_sd_up     4   NA       4       4     4        1
#> 10 diff_sd_down NaN   NA      NA     Inf  -Inf        2
#> 11 diff_n         1   NA       1       1     1        1
#> 12 diff_n_up      1   NA       1       1     1        1
#> 13 diff_n_down   -1   NA      -1      -1    -1        1