Detect duplicate values

For every value in a vector or data frame, duplicate_detect() tests whether there is at least one identical value. Test results are presented next to every value.

By default, only numeric columns and string columns coercible to numeric are tested (if x is a data frame). Any other columns are silently dropped.

This function is a blunt tool designed for initial data checking. Don't put too much weight on its results.

For summary statistics, call audit() on the results.

Usage

duplicate_detect(x, numeric_only = TRUE, colname_end = "dup")

Arguments

x

Vector or data frame.

numeric_only

Boolean. If TRUE (the default) and if x is a data frame, the function will only test numeric columns and string columns coercible to numeric. Note: Be careful when setting it to FALSE. This can lead to all kinds of coercion issues.

colname_end

String. Name ending of the Boolean test result columns. Default is "dup".

Value

A tibble (data frame) —

If x is a vector, there are two columns: the input value and the Boolean has_duplicates.

If x is a data frame, the output tibble has (some of) the columns from x, and to each of these columns' right, the corresponding Boolean column with an index value.

The tibble has the scr_dup_detect class, which is recognized by the audit() generic.

Details

This function is not very informative with many input values that only have a few characters each. Many of them may have duplicates just by chance. For example, in R's built-in iris data set, 99% of values have duplicates.

In general, the fewer values and the more characters per value there are, the more significant duplicate_detect()'s results will be.

Summaries with audit()

There is an S3 method for the audit() generic, so you can call audit() following duplicate_detect(). It returns a tibble with these columns ---

variable: The original data frame's variables with at least one "duplicated" value: one that has at least one duplicate anywhere else in the data frame. For a vector, x.

n_duplicated: Number of "duplicated" values of that variable: those that have at least one duplicate anywhere in the data frame.

dup_rate: Rate of "duplicated" values of that variable.

The final row, .total, summarizes across all other rows: It adds up the n_duplicated and n_total columns, and calculates the average of the dup_rate column.

See also

Examples

# Find duplicate values in a data frame...
duplicate_detect(x = pigs4)
#> # A tibble: 5 × 4
#>   snout snout_dup  tail tail_dup
#>   <dbl> <lgl>     <dbl> <lgl>   
#> 1  4.74 FALSE      6.89 FALSE   
#> 2  8.13 TRUE       7.33 FALSE   
#> 3  4.22 TRUE       6.10 FALSE   
#> 4  4.22 TRUE       7.57 FALSE   
#> 5  5.18 FALSE      8.13 TRUE    

# ...or in a single vector:
duplicate_detect(x = pigs4$snout)
#> # A tibble: 5 × 2
#>   value has_duplicates
#>   <dbl> <lgl>         
#> 1  4.74 FALSE         
#> 2  8.13 FALSE         
#> 3  4.22 TRUE          
#> 4  4.22 TRUE          
#> 5  5.18 FALSE         

# Summary statistics with `audit()`:
pigs4 %>%
  duplicate_detect() %>%
  audit()
#> # A tibble: 3 × 4
#>   variable n_duplicated n_total dup_rate
#>   <chr>           <int>   <int>    <dbl>
#> 1 snout               3       5      0.6
#> 2 tail                1       5      0.2
#> 3 .total              4      10      0.4

# If there are many values and/or few
# characters per value, `duplicate_detect()`
# can be misleading:
iris %>%
  duplicate_detect()
#> # A tibble: 150 × 8
#>    Sepal.Length Sepal.Length_dup Sepal…¹ Sepal…² Petal…³ Petal…⁴ Petal…⁵ Petal…⁶
#>           <dbl> <lgl>              <dbl> <lgl>     <dbl> <lgl>     <dbl> <lgl>  
#>  1          5.1 TRUE                 3.5 TRUE        1.4 TRUE        0.2 TRUE   
#>  2          4.9 TRUE                 3   TRUE        1.4 TRUE        0.2 TRUE   
#>  3          4.7 TRUE                 3.2 TRUE        1.3 TRUE        0.2 TRUE   
#>  4          4.6 TRUE                 3.1 TRUE        1.5 TRUE        0.2 TRUE   
#>  5          5   TRUE                 3.6 TRUE        1.4 TRUE        0.2 TRUE   
#>  6          5.4 TRUE                 3.9 TRUE        1.7 TRUE        0.4 TRUE   
#>  7          4.6 TRUE                 3.4 TRUE        1.4 TRUE        0.3 TRUE   
#>  8          5   TRUE                 3.4 TRUE        1.5 TRUE        0.2 TRUE   
#>  9          4.4 TRUE                 2.9 TRUE        1.4 TRUE        0.2 TRUE   
#> 10          4.9 TRUE                 3.1 TRUE        1.5 TRUE        0.1 TRUE   
#> # … with 140 more rows, and abbreviated variable names ¹​Sepal.Width,
#> #   ²​Sepal.Width_dup, ³​Petal.Length, ⁴​Petal.Length_dup, ⁵​Petal.Width,
#> #   ⁶​Petal.Width_dup

iris %>%
  duplicate_detect() %>%
  audit()
#> # A tibble: 5 × 4
#>   variable     n_duplicated n_total dup_rate
#>   <chr>               <int>   <int>    <dbl>
#> 1 Petal.Length          150     150    1    
#> 2 Petal.Width           148     150    0.987
#> 3 Sepal.Length          144     150    0.96 
#> 4 Sepal.Width           150     150    1    
#> 5 .total                592     600    0.987