Create new *_map_total_n() functions

function_map_total_n() is the engine that powers functions such as grim_map_total_n(). It creates new, "manufactured" functions for consistency tests. The new functions take reported summary statistics such as means and apply those tests in cases where only a total sample size is known, not group sizes.

This works by making disperse_total() create multiple pairs of hypothetical group sizes, all of which add up to the reported total. There need to be exactly two groups.

If functions created this way are exported from other packages, they should be written as if they were created with purrr adverbs; see explanations there and examples at vignette("consistency-tests"), section Creating mappers with function_map().

Usage

function_map_total_n(
  .fun,
  .reported,
  .name_test,
  .name_class = NULL,
  .dispersion = 0:5,
  .n_min = 1,
  .n_max = NULL,
  .constant = NULL,
  .constant_index = NULL
)

Arguments

.fun

Function such as grim_map: It will be used to test columns in a data frame for consistency. Test results are Boolean and need to be contained in a column called consistency that is added to the input data frame. This modified data frame is then returned by .fun.

.reported

String. Names of the columns containing group-specific statistics that were reported alongside the total sample size(s). They will be tested for consistency with the hypothetical group sizes. Examples are "x" for GRIM and c("x", "sd") for DEBIT. In the data frame with reported group statistics that the manufactured function takes as an input, each will need to fan out like "x1", "x2", "sd1", and "sd2".

.name_test

String (length 1). The name of the consistency test, such as "GRIM", to be optionally shown in a message when using the manufactured function.

.name_class

String. If specified, the tibbles returned by the manufactured function will inherit this string as an S3 class. Default is NULL, i.e., no extra class.

.dispersion, .n_min, .n_max, .constant, .constant_index

Arguments passed down to disperse_total(), using defaults from there.

Value

A function such as these:

Manufactured function	Reported statistics	Test vignette

grim_map_total_n()	"x"	vignette("grim")

grimmer_map_total_n()	"x", "sd"	vignette("grimmer")

debit_map_total_n()	"x", "sd"	vignette("debit")

The factory-made function will also have dots, ..., to pass arguments down to .fun, i.e., the basic mapper function.

Details

This function is a so-called function factory: It produces other functions, such as grim_map_total_n(). More specifically, it is a function operator (a.k.a. decorator) because it also takes functions as inputs, such as grim_map(). See Wickham (2019, ch. 10-11).

Conventions

The name of a function manufactured with function_map_total_n() should mechanically follow from that of the input function. For example, grim_map_total_n() derives from grim_map(). This pattern fits best if the input function itself is named after the test it performs on a data frame, followed by _map: grim_map() applies GRIM, debit_map() applies DEBIT, etc.

Much the same is true for the classes of data frames returned by the manufactured function via the .name_class argument of function_map_total_n(). It should be the function's own name preceded by the name of the package that contains it or by an acronym of that package's name. In this way, existing classes are scr_grim_map_total_n and scr_debit_map_total_n.

References

Bauer, P. J., & Francis, G. (2021). Expression of Concern: Is It Light or Dark? Recalling Moral Behavior Changes Perception of Brightness. Psychological Science, 32(12), 2042–2043. https://journals.sagepub.com/doi/10.1177/09567976211058727

Wickham, H. (2019). Advanced R (Second Edition). CRC Press/Taylor and Francis Group. https://adv-r.hadley.nz/index.html

See also

Examples

# Function definition of `grim_map_total_n()`:
grim_map_total_n <- function_map_total_n(
  .fun = grim_map,
  .reported = "x",
  .name_test = "GRIM",
  .name_class = "scr_grim_map_total_n"
)

# Case study of SCHLIM, a new consistency test --------------

# (Note: This is a mock test without any real significance.
# Its only purpose is to show the minimal steps necessary
# for implementing a serious consistency test, and to use
# it as a starting point for `function_map_total_n()`.)

# The "SCHLIM test" is analogous to GRIM as implemented
# in scrutiny. This is also true for the function names.
# Note that the analogue to `schlim_scalar()`, a function
# called `grim_scalar()`, is not exported from scrutiny,
# but used internally for `grim()`, `grim_map()`, and,
# indirectly, `grim_map_total_n()`.

# Basic test implementation:
schlim_scalar <- function(y, n) {
  (y / 3) > n
}

# This step is not needed below, but
# included for completeness:
schlim <- Vectorize(schlim_scalar)

# This will be the input function for
# `function_map_total_n()`:
schlim_map <- function_map(
  .fun = schlim_scalar,
  .reported = c("y", "n"),
  .name_test = "SCHLIM"
)

# Fire up the function factory:
schlim_map_total_n <- function_map_total_n(
  .fun = schlim_map,
  .reported = "y",
  .name_test = "SCHLIM",
)

# Create some example data:
df1 <- tibble::tibble(
  y1 = 16:25,
  y2 = 26:35,
  n  = 12:21
)
df1
#> # A tibble: 10 × 3
#>       y1    y2     n
#>    <int> <int> <int>
#>  1    16    26    12
#>  2    17    27    13
#>  3    18    28    14
#>  4    19    29    15
#>  5    20    30    16
#>  6    21    31    17
#>  7    22    32    18
#>  8    23    33    19
#>  9    24    34    20
#> 10    25    35    21

# Call the manufactured function:
out <- schlim_map_total_n(df1)
out
#> # A tibble: 240 × 7
#>        y     n n_change consistency both_consistent  case dir  
#>    <int> <dbl>    <dbl> <lgl>       <lgl>           <int> <chr>
#>  1    16     6        0 FALSE       FALSE               1 forth
#>  2    26     6        0 TRUE        FALSE               1 forth
#>  3    16     5       -1 TRUE        TRUE                1 forth
#>  4    26     7        1 TRUE        TRUE                1 forth
#>  5    16     4       -2 TRUE        TRUE                1 forth
#>  6    26     8        2 TRUE        TRUE                1 forth
#>  7    16     3       -3 TRUE        FALSE               1 forth
#>  8    26     9        3 FALSE       FALSE               1 forth
#>  9    16     2       -4 TRUE        FALSE               1 forth
#> 10    26    10        4 FALSE       FALSE               1 forth
#> # … with 230 more rows

# Summarize the results:
audit_total_n(out)
#> # A tibble: 10 × 8
#>       y1    y2     n hits_total hits_forth hits_back scenarios_total hit_rate
#>    <int> <int> <dbl>      <dbl>      <dbl>     <dbl>           <dbl>    <dbl>
#>  1    16    26    12          2          2         0              12   0.167 
#>  2    17    27    13          1          1         0              12   0.0833
#>  3    18    28    14          1          1         0              12   0.0833
#>  4    19    29    15          1          1         0              12   0.0833
#>  5    20    30    16          0          0         0              12   0     
#>  6    21    31    17          0          0         0              12   0     
#>  7    22    32    18          0          0         0              12   0     
#>  8    23    33    19          0          0         0              12   0     
#>  9    24    34    20          0          0         0              12   0     
#> 10    25    35    21          0          0         0              12   0